subscribe: return error if both QoS bits are set
The upper 6 bits of the Requested QoS byte are not used in the current version
of the protocol. They are reserved for future use. The Server MUST treat a
SUBSCRIBE packet as malformed and close the Network Connection if any of
Reserved bits in the payload are non-zero, or QoS is not 0,1 or 2
[MQTT-3-8.3-4].

http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc398718066